requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0
boto3==1.34.0
structlog==23.2.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2
python-dotenv==1.0.0
gunicorn==21.2.0
PyJWT==2.8.0
cryptography==41.0.7
mediapipe==0.10.7
opencv-python==4.8.1.78
numpy==1.24.3
